+++ activerecord/lib/active_record/railtie.rb
@@ -37,6 +37,7 @@ module ActiveRecord
     config.active_record.cache_query_log_tags = false
     config.active_record.query_log_tags_prepend_comment = false
     config.active_record.raise_on_assign_to_attr_readonly = false
+    config.active_record.has_many_strict_replace = false
     config.active_record.belongs_to_required_validates_foreign_key = true
     config.active_record.generate_secure_token_on = :create
     config.active_record.use_legacy_signed_id_verifier = :generate_and_verify